refactored retry mechanic into a mixin class
