+++ src/sentry/models/project.py
@@ -35,7 +35,7 @@ from sentry.utils.colors import get_hashed_color
 from sentry.utils.http import absolute_uri
 from sentry.utils.integrationdocs import integration_doc_exists
 from sentry.utils.retries import TimedRetryPolicy
-from sentry.utils.snowflake import snowflake_id_generation
+from sentry.utils.snowflake import SnowflakeIdMixin
 
 if TYPE_CHECKING:
     from sentry.models import User
@@ -99,7 +99,7 @@ class ProjectManager(BaseManager):
         return sorted(project_list, key=lambda x: x.name.lower())
 
 
-class Project(Model, PendingDeletionMixin):
+class Project(Model, PendingDeletionMixin, SnowflakeIdMixin):
     from sentry.models.projectteam import ProjectTeam
 
     """
@@ -143,8 +143,6 @@ class Project(Model, PendingDeletionMixin):
     objects = ProjectManager(cache_fields=["pk"])
     platform = models.CharField(max_length=64, null=True)
 
-    snowflake_redis_key = "project_snowflake_key"
-
     class Meta:
         app_label = "sentry"
         db_table = "sentry_project"
@@ -167,9 +165,7 @@ class Project(Model, PendingDeletionMixin):
             span.set_data("project_slug", self.slug)
             return Counter.increment(self)
 
-    def save(self, snowflake_retry_counter=0, *args, **kwargs):
-        if not self.id:
-            self.id = snowflake_id_generation(self.snowflake_redis_key)
+    def save(self, *args, **kwargs):
         if not self.slug:
             lock = locks.get("slug:project", duration=5)
             with TimedRetryPolicy(10)(lock.acquire):
@@ -180,15 +176,12 @@ class Project(Model, PendingDeletionMixin):
                     reserved=RESERVED_PROJECT_SLUGS,
                     max_length=50,
                 )
-        try:
-            with transaction.atomic():
-                super().save(*args, **kwargs)
-        except IntegrityError:
-            if snowflake_retry_counter > settings.MAX_REDIS_SNOWFLAKE_RETY_COUNTER:
-                raise Exception("Max allowed ID retry reached. Please try again in a second")
-            snowflake_retry_counter += 1
-            self.id = None
-            self.save(snowflake_retry_counter, *args, **kwargs)
+
+        snowflake_redis_key = "project_snowflake_key"
+        self.save_with_snowflake_id(
+            snowflake_redis_key, lambda: super(Project, self).save(*args, **kwargs)
+        )
+
         self.update_rev_for_option()
 
     def get_absolute_url(self, params=None):
